requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests>=2.0
mercadopago
Flask-SQLAlchemy==3.1.1
alembic>=1.13.0
psycopg2-binary>=2.9.9
